+++ src/chameleon/zpt/loader.py
@@ -10,6 +10,13 @@ class TemplateLoader(BaseLoader):
 
     default_format = "xml"
 
+    def __init__(self, **kwargs):
+        formats = kwargs.pop('formats', None)
+        if formats is not None:
+            self.formats = formats
+
+        super(TemplateLoader, self).__init__(**kwargs)
+
     def load(self, filename, format=None):
         """Load and return a template file.